requirements.txt
fastapi==0.115.12
httpx==0.28.1
setuptools==78.1.0
uvicorn==0.34.0
pydantic-settings==2.8.1
PyYAML==6.0.2
python-json-logger==3.3.0